refactor out of check.ts into run.ts
